Add wheel and ball animations with result evaluation and payout system
- Added spinning animations for wheel (American & European variants)
- Implemented ball movement: circling, bouncing, and entering selected slot
- Introduced event triggers: OnBallStopped, OnBetResult..
- Calculated winning bets and applied payouts to player balance
- Set up base systems for both American and European roulette
-Basic game loop within roulettemanager, states..

+++ Assets/Scripts/Gameplay/Betting/Data/BetAction.cs
@@ -3,6 +3,6 @@ namespace Gameplay.Betting.Data
     public struct BetAction
     {
         public BetAnchor Anchor;
-        public int Value;
+        public long Value;
     }
 }
